Fix tests failing due to lack of backend
Several tests failed when run on their own. In some cases they could be
fixed with a basic mock, in another I moved it to an integration
test module.
