+++ docker/cla_public.ini
@@ -7,7 +7,6 @@ chmod-socket = 666
 chown-socket = www-data
 master = true
 enable-threads = true
-plugin = python
 processes = 2
 chdir = /home/app/flask
 module = cla_public.wsgi
